Xuất file excel

+++ SourceCode/EmployeeTracking.Admin/Filters/DeleteFileAttribute.cs
@@ -11,8 +11,11 @@ namespace EmployeeTracking.Admin.Filters
             //convert the current filter context to file and get the file path
             string filePath = (filterContext.Result as FilePathResult).FileName;
 
-            //delete the file after download
-            System.IO.File.Delete(filePath);
+            if (System.IO.File.Exists(filePath))
+            {
+                //delete the file after download
+                System.IO.File.Delete(filePath);
+            }
         }
     }
 }
